dhclient: overwrite DUID in lease file when a new duid is specified
We have just introduced the brand new dhcp-duid property to allow to
take control of the DUID in DHCPv6: use the specified value when
provided.
Note that the DUID in the lease file is still the default option when no
duid value is provided both at connection and system level.
If a default duid is globally provided, the one in the connection lease
file could still be preferred setting the "dhcp-duid" property to the
special "lease" value.